FEAT: Change api to create current user memo
