+++ tests/test_api/test_routes/test_memos.py
@@ -1,11 +1,11 @@
 from fastapi import status
+from fastapi.applications import FastAPI
 from fastapi.testclient import TestClient
 
 
-class TestGetMemo:
-    def test_routes_exist(
-        self,
-        client: TestClient
-    ) -> None:
-        res = client.get("/api/memos")
-        assert res.status_code == status.HTTP_200_OK
+def test_user_can_get_own_memos(
+    app: FastAPI,
+    authorized_client: TestClient
+) -> None:
+    res = authorized_client.get(app.url_path_for("memos:get-own-memos"))
+    assert res.status_code == status.HTTP_200_OK
